requirements.txt
aiofiles==22.1.0
asyncpg==0.26.0
cffi==1.15.1
cryptography==38.0.1
greenlet==1.1.3.post0
httptools==0.5.0
multidict==6.0.2
psycopg2-binary==2.9.4
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.6.0
PyYAML==6.0
sanic==22.9.0
sanic-ext==22.9.0
sanic-routing==22.8.0
SQLAlchemy==1.4.42
ujson==5.5.0
uvloop==0.17.0
websockets==10.3
